+++ Back-end/src/main/java/barbershop_vk/entity/User.java
@@ -4,7 +4,7 @@ import java.util.Objects;
 
 public class User {
 
-    private int id;
+    private Long id;
     private String name;
     private String email;
     private String password;
@@ -13,7 +13,7 @@ public class User {
     public User() {
     }
 
-    public User(int id, String name, String email, String password, String telephone) {
+    public User(Long id, String name, String email, String password, String telephone) {
         this.id = id;
         this.name = name;
         this.email = email;
@@ -29,11 +29,11 @@ public class User {
         this.email = email;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
